refactor(api): Extract pipeline_types.rs (3 DTOs)
- Create pipeline_types.rs with EnhancedPipelineStatusResponse, PipelineMessageResponse, CancelPipelineResponse
- Add 3 unit tests for response serialization
- pipeline.rs: 238 -> 90 lines (-62% reduction)
- Total API lib tests: 338
