+++ edgequake/crates/edgequake-api/src/handlers/mod.rs
@@ -21,6 +21,7 @@ pub mod metrics;
 pub mod ollama;
 pub mod ollama_types;
 pub mod pipeline;
+pub mod pipeline_types;
 pub mod query;
 pub mod query_types;
 pub mod relationships;
@@ -53,6 +54,7 @@ pub use metrics::*;
 pub use ollama::*;
 pub use ollama_types::*;
 pub use pipeline::*;
+pub use pipeline_types::*;
 pub use query::*;
 pub use query_types::*;
 pub use relationships::*;
